ext_framebuffer_multisample: fix int-to-float narrowing in draw-buffers-common
Newer versions of Apple Clang (shipped with Xcode on arm64 macOS)
treat -Wc++11-narrowing as an error by default, unlike GCC and
older Clang which only emit a warning. This causes a build failure
where the int variable 'pattern_width' is used to brace-initialize
a float array, which is a narrowing conversion forbidden by C++11.

Fix by explicitly casting pattern_width to float in the initializer
list.

Fixes: 9f5d8b334 ("msaa: Add files containing the shared code to test multiple draw buffers")
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1089>

+++ tests/spec/ext_framebuffer_multisample/draw-buffers-common.cpp
@@ -353,8 +353,8 @@ draw_pattern(bool sample_alpha_to_coverage,
 		float vertices[4][2] = {
 		{ 0.0f, 0.0f + i * (pattern_height / num_rects) },
 		{ 0.0f, (i + 1.0f) * (pattern_height / num_rects) },
-		{ pattern_width, (i + 1.0f) * (pattern_height / num_rects) },
-		{ pattern_width, 0.0f + i * (pattern_height / num_rects) } };
+		{ static_cast<float>(pattern_width), (i + 1.0f) * (pattern_height / num_rects) },
+		{ static_cast<float>(pattern_width), 0.0f + i * (pattern_height / num_rects) } };
 
 		glVertexAttribPointer(0, 2, GL_FLOAT, GL_FALSE,
 				      sizeof(vertices[0]),
